+++ ACELEROMETRO_MMA8451Q/sdk_hal/sdk_hal_i2c0.h
@@ -79,9 +79,9 @@ status_t i2c0MasterInit(uint32_t baud_rate);
  * 		kStatus_Fail
  * @endcode
  */
-status_t i2c0MasterReadByte(uint8_t *data, uint8_t device_address, int8_t memory_address);
+status_t i2c0MasterReadByte(uint16_t *data, uint8_t device_address, int8_t memory_address);
 
-status_t i2c0MasterWriteByte(uint8_t *data, uint8_t device_address, int8_t reg_addr, uint8_t value);
+status_t i2c0MasterWriteByte(uint16_t *data, uint8_t device_address, int8_t reg_addr, uint8_t value);
